Ported and merged in ratbox's revisions 25354 and 25364
Fixes some problems with the kqueue event engine,
including a bug when flushing the event queue a
closed FD would cause the kernel to ignore any events
for FDs after it resulting in ignoring read and write for
certain sockets.

For more information see the comments in libratbox's
kqueue engine.